retl/requirements.txt
pandas==2.3.0
google-cloud-storage==3.1.0
google-cloud-bigquery==3.34.0
db-dtypes==1.4.3
pydantic==2.11.5
